backend/requirements.dev.txt
argon2-cffi==21.3.0
black==22.8.0
Django==3.2.0
django-celery-beat==2.3.0
django-compressor==4.1
django-cors-headers==3.13.0
django-enum-choices==2.1.4
django-extensions==3.2.1
django-filter==22.1
django-ipware==4.0.2
django-jazzmin==2.5.0
django-redis==5.2.0
django-simple-history==3.1.1
djangorestframework-simplejwt==5.2.0
flake8-isort==4.2.0
gunicorn==20.1.0
ipython==8.5.0
pre-commit==2.20.0
psycopg2-binary==2.9.3
pylint-celery==0.3
pylint-django==2.5.3
python-dotenv==0.21.0
